chore: moved pagination class into a separate file. Updated files that used it and re-formatted them clean them up. (last commit didn't contain changes)
