+++ src/app/namespace/namespace-manager/namespace-manager.component.ts
@@ -1,59 +1,59 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
-import { NamespaceServiceService } from "../../../api";
-import { Pagination } from "../../workflow-template/workflow-template-view/workflow-template-view.component";
-import { FormControl } from "@angular/forms";
-import { debounceTime, distinctUntilChanged } from "rxjs/operators";
+import { NamespaceServiceService } from '../../../api';
+import { FormControl } from '@angular/forms';
+import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { Pagination } from '../../requests/pagination';
 
 @Component({
-  selector: 'app-namespace-manager',
-  templateUrl: './namespace-manager.component.html',
-  styleUrls: ['./namespace-manager.component.scss']
+    selector: 'app-namespace-manager',
+    templateUrl: './namespace-manager.component.html',
+    styleUrls: ['./namespace-manager.component.scss']
 })
 export class NamespaceManagerComponent implements OnInit {
-  @Output() namespaceSelected = new EventEmitter<string>();
-  @Output() newNamespaceClick = new EventEmitter();
+    @Output() namespaceSelected = new EventEmitter<string>();
+    @Output() newNamespaceClick = new EventEmitter();
 
-  queryInput: FormControl;
-  pagination = new Pagination();
+    queryInput: FormControl;
+    pagination = new Pagination();
 
-  namespaces = [];
+    namespaces = [];
 
-  constructor(private namespaceServiceService: NamespaceServiceService) { }
+    constructor(private namespaceServiceService: NamespaceServiceService) {
+    }
 
-  ngOnInit() {
-    this.queryInput = new FormControl('');
-    this.queryInput.valueChanges.
-        pipe(
+    ngOnInit() {
+        this.queryInput = new FormControl('');
+        this.queryInput.valueChanges.pipe(
             debounceTime(400),
             distinctUntilChanged(),
         )
-        .subscribe(res => {
-            this.getNamespaces();
-        });
+            .subscribe(res => {
+                this.getNamespaces();
+            });
 
-    this.getNamespaces();
-  }
+        this.getNamespaces();
+    }
 
-  getNamespaces() {
-    const query = this.queryInput.value;
-    const pageSize = this.pagination.pageSize;
-    const page = this.pagination.page + 1;
+    getNamespaces() {
+        const query = this.queryInput.value;
+        const pageSize = this.pagination.pageSize;
+        const page = this.pagination.page + 1;
 
-    this.namespaceServiceService.listNamespaces(pageSize, page, query)
-        .subscribe(res => {
-            if(!res.namespaces) {
-                this.namespaces = [];
-            } else {
-                this.namespaces = res.namespaces;
-            }
-        });
-  }
+        this.namespaceServiceService.listNamespaces(pageSize, page, query)
+            .subscribe(res => {
+                if (!res.namespaces) {
+                    this.namespaces = [];
+                } else {
+                    this.namespaces = res.namespaces;
+                }
+            });
+    }
 
-  onNamespaceClick(namespace: string) {
-    this.namespaceSelected.emit(namespace);
-  }
+    onNamespaceClick(namespace: string) {
+        this.namespaceSelected.emit(namespace);
+    }
 
-  onNewNamespaceClick() {
-      this.newNamespaceClick.emit();
-  }
+    onNewNamespaceClick() {
+        this.newNamespaceClick.emit();
+    }
 }
