+++ src/app/requests/pagination.ts
@@ -1,0 +1,4 @@
+export class Pagination {
+    page = 0;
+    pageSize = 15;
+}
